fix(#45): add StereoCompose to the Android target; run the tests in CI
The Android build has its own curated source list and compiles LifLoader.cpp
directly, so factoring ComposeSbs + EstimateAutoConvergence out into
StereoCompose.cpp broke its link with two undefined symbols. Desktop CI
could not catch it -- every other target builds the whole media directory.
Add the missing TU.

Also wire the unit tests into macOS CI. MP_BUILD_TESTS defaults OFF and no
workflow ever set it or ran ctest, so lif_roundtrip_test had never executed
in CI and the three new suites would not have either. The step
re-configures the tree build_macos.sh already populated, so it is
incremental and re-fetches nothing; stereo_sweep skips cleanly when the
generated corpus is absent, which it is on CI.

And cast AVFrameSideData::size to size_t before comparing: the field
changed int -> size_t across ffmpeg versions, so the bare comparison is a
sign-mismatch on one side or the other.

+++ src/media/VideoStereoProbe.cpp
@@ -46,7 +46,9 @@ constexpr int kDarkMean = 12;    // frames dimmer than this are skipped outright
 // 6.1 are both major 60 — and CI still builds against ffmpeg 4.4 on Ubuntu 22.04.
 bool ReadStereo3D(const AVFrame* f, StereoLayout& layout, bool& invert, const char*& why) {
     const AVFrameSideData* sd = av_frame_get_side_data(f, AV_FRAME_DATA_STEREO3D);
-    if (!sd || sd->size < (int)sizeof(AVStereo3D)) return false;
+    // AVFrameSideData::size changed int -> size_t across ffmpeg versions; cast both to
+    // size_t so the comparison neither warns nor sign-flips on either.
+    if (!sd || (size_t)sd->size < sizeof(AVStereo3D)) return false;
     const AVStereo3D* s3 = reinterpret_cast<const AVStereo3D*>(sd->data);
 
     invert = (s3->flags & AV_STEREO3D_FLAG_INVERT) != 0;
